add ES3 test for R.keys
