+++ package.json
@@ -42,7 +42,7 @@
     "postcoverage": "npm run posttest",
     "lint": "eslint scripts/bookmarklet scripts/build src/*.js src/internal/*.js test/*.js test/**/*.js lib/sauce/*.js lib/bench/*.js",
     "pretest": "npm install && npm run lint && npm run clean && npm run build",
-    "test": "mocha --reporter spec",
+    "test": "mocha --reporter spec && mocha test/internal",
     "posttest": "git checkout -- dist",
     "prebrowser_test": "npm run pretest",
     "browser_test": "testem ci",
