+++ test/internal/_keysFallback.js
@@ -1,0 +1,4 @@
+var _keysFallback = require('../../src/internal/_keysFallback');
+var keys = require('../shared/keys');
+
+keys('keysFallback', _keysFallback);
